fix pexpect usage
pexpect wasn't finding the right prompt match and the kernel would then
timeout on the jupyter side and get killed.

Use `bash --login` and assume that it's immediately there instead of
waiting for the first prompt.

Pass the unformatted `prompt_change` to the parent class as it's going
to take care of the formatting by itself. This will make it more
resitant to changes in pexpect.